+++ force-app/main/default/lwc/serviceScheduleCreator/navigationItem.js
