+++ force-app/main/default/lwc/serviceScheduleCreator/navigationItems.js
