+++ force-app/main/default/lwc/serviceScheduleCreator/progressStep.js
