+++ force-app/main/default/lwc/serviceScheduleCreator/progressSteps.js
