+++ force-app/main/default/lwc/serviceScheduleCreator/serviceScheduleCreator.html
@@ -19,12 +19,62 @@
             <c-section title={labels.reviewSchedule}></c-section>
         </template>
         <div slot="footer">
-            <c-progress-navigator
-                if:true={currentStep}
-                steps={steps}
-                current-step={currentStep}
-                onnavigate={handleNavigation}
-            ></c-progress-navigator>
+            <lightning-layout horizontal-align="end">
+                <lightning-layout-item
+                    padding="around-small"
+                    size="3"
+                    class="slds-text-align_left"
+                >
+                    <lightning-button
+                        if:true={currentStep.previous}
+                        name="previous"
+                        variant={currentStep.previous.variant}
+                        label={currentStep.previous.label}
+                        title={currentStep.previous.label}
+                        onclick={handlePrevious}
+                    ></lightning-button>
+                </lightning-layout-item>
+
+                <lightning-layout-item
+                    padding="around-small"
+                    size="6"
+                    class="slds-text-align_center"
+                >
+                    <lightning-progress-indicator
+                        current-step={currentStep.value}
+                        type="base"
+                        variant="base"
+                    >
+                        <template for:each={steps} for:item="step">
+                            <lightning-progress-step
+                                key={step.value}
+                                label={step.label}
+                                value={step.value}
+                            ></lightning-progress-step>
+                        </template>
+                    </lightning-progress-indicator>
+                </lightning-layout-item>
+
+                <lightning-layout-item padding="around-small" size="3">
+                    <lightning-button
+                        if:true={currentStep.next}
+                        name="next"
+                        variant={currentStep.next.variant}
+                        label={currentStep.next.label}
+                        title={currentStep.next.label}
+                        onclick={handleNext}
+                        class="slds-var-p-right_small"
+                    ></lightning-button>
+                    <lightning-button
+                        if:true={currentStep.finish}
+                        name="finish"
+                        variant={currentStep.finish.variant}
+                        label={currentStep.finish.label}
+                        title={currentStep.finish.label}
+                        onclick={handleFinish}
+                    ></lightning-button>
+                </lightning-layout-item>
+            </lightning-layout>
         </div>
     </c-modal>
 </template>
